requirements.txt
Flask==2.3.3
python-docx==0.8.11
reportlab==4.0.4
openai==1.3.0
spacy==3.7.2